+++ changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.3.7] Unreleased Github Repo [develop]
+### Fix
+- paste_buff method in BleDevice
+### Added
+- @code decorator for Device classes to declare and call functions in devices
 ## [0.3.6] - 2022-03-25
 ### Fix
 - fixed missing dependencies in `setup.py` (@BradenM)